+++ src/presenter/components/index.ts
@@ -1,14 +1,14 @@
-export { default as Badge } from "@components/Badge";
-export { default as Box } from "@components/Box";
-export { default as Breadcrumb } from "@components/Breadcrumb";
-export { default as Button } from "@components/Button";
-export { default as ContactTable } from "@components/ContactTable";
-export { default as Container } from "@components/Container";
-export { default as DropdownMenu } from "@components/DropdownMenu";
-export { default as Flex } from "@components/Flex";
-export { default as Form } from "@components/Form";
-export { default as Header } from "@components/Header";
-export { default as Input } from "@components/Input";
-export { default as Loader } from "@components/Loader";
-export { default as Modal } from "@components/Modal";
-export { default as Select } from "@components/Select";
+export { default as Badge } from "@/presenter/components/Badge";
+export { default as Box } from "@/presenter/components/Box";
+export { default as Breadcrumb } from "@/presenter/components/Breadcrumb";
+export { default as Button } from "@/presenter/components/Button";
+export { default as ContactTable } from "@/presenter/components/ContactTable";
+export { default as Container } from "@/presenter/components/Container";
+export { default as DropdownMenu } from "@/presenter/components/DropdownMenu";
+export { default as Flex } from "@/presenter/components/Flex";
+export { default as Form } from "@/presenter/components/Form";
+export { default as Header } from "@/presenter/components/Header";
+export { default as Input } from "@/presenter/components/Input";
+export { default as Loader } from "@/presenter/components/Loader";
+export { default as Modal } from "@/presenter/components/Modal";
+export { default as Select } from "@/presenter/components/Select";
